refactor(ui): reuse the kit clipboard and attachment src hooks
