+++ apps/docs/components/pages/examples/use-attachment-src.ts
@@ -1,40 +1,5 @@
 "use client";
 
-import { useEffect, useState } from "react";
-import { useAuiState } from "@assistant-ui/react";
-import { useShallow } from "zustand/shallow";
+import { useAttachmentSrc } from "@assistant-ui/ui/hooks/use-attachment-src";
 
-const useFileSrc = (file: File | undefined) => {
-  const [src, setSrc] = useState<string | undefined>(undefined);
-
-  useEffect(() => {
-    if (!file) {
-      setSrc(undefined);
-      return;
-    }
-
-    const objectUrl = URL.createObjectURL(file);
-    setSrc(objectUrl);
-
-    return () => {
-      URL.revokeObjectURL(objectUrl);
-    };
-  }, [file]);
-
-  return src;
-};
-
-export const useAttachmentSrc = () => {
-  const { file, src } = useAuiState(
-    useShallow((s): { file?: File; src?: string } => {
-      if (s.attachment.type !== "image") return {};
-      if (s.attachment.file) return { file: s.attachment.file };
-      const src = s.attachment.content?.filter((c) => c.type === "image")[0]
-        ?.image;
-      if (!src) return {};
-      return { src };
-    }),
-  );
-
-  return useFileSrc(file) ?? src;
-};
+export { useAttachmentSrc };
